server doesnt start when initialized; refactored NetworkInterfaceMaster calbak
the server address is set only when it is found, then the callback refreshes the server which then updates the prompt through the passed Facilitator.

old issue still persists:
should cover most of connections; what if the user is connected through usb sharing(should be included in eth), but wants to connect with some other device via wifi?